+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed client_name to user_agent_suffix
+- Use package name as prefix instead of ExApiClient.
+- set Mix.Project.config[:app] as a default value for user_agent_suffix
 
 ## [1.0.0] - 2017-10-17
 ### Added